chore: 초기 설정
refactor: 필요없는 파일 삭제

필요없는 파일 삭제

chore: npm i react-router-dom

npm i react-router-dom

chore: npm i axios

npm i axios

chore: npm i styled-components

npm i styled-components

chore: npm i eslint, prettier, airbnb, husky

npm i eslint, prettier, airbnb, husky
