+++ src/App.jsx
@@ -1,0 +1,5 @@
+function App() {
+  return <div></div>;
+}
+
+export default App;
